Initial commit
csi-connector is intended to be a general purpose golang library
that can be used to connect CSI Volumes to Nodes.  We're starting
with iSCSI, and will also add other modules for things like FC
or ISER as needed.

The goal is to make this as simple, yet as complete as possible
using only std golang imports.