Add some unit tests
This changes adds some unit tests (including fixes for the session
parser that the unit tested detected a bug in).

This also prompted the refactoring of how we do exec.Command, we use a
variable to set that now so we can easily use the built in test-helper
provided in exec.Command to fake things out.

Test_sessionExists is a nice example of this where we test multiple
function calls with a single fake of exec.Command